more user friendly intro message

+++ setup.sh
@@ -1,8 +1,22 @@
 #!/bin/bash
-echo "This script will replace your zsh, vimrc files with this package's version.  "
-echo "you should have this project checkedout in $HOME/bin, otherwise it'll fail/do something stupid"
-echo "hit Ctrl+C to abort"
-sleep 5
+#############################################################################
+# Automated installation script
+#  - This script will set up common directories, dot files and other system
+#    settings
+#############################################################################
+echo "This script will automatically update your system to use common directories, "
+echo "various dotfiles (zsh, vimrc, etc) and other assorted goodies. "
+echo " "
+
+# Ask the user to confirm before proceeding
+read -p "Do you wish to continue? (y|n) "
+if [[ ! $REPLY =~ ^[Yy]$ ]]
+then
+    exit 1
+fi
+
+exit
+
 rm -frv $HOME/{.zshrc,.aliases,.vim,.vimrc}
 cd $HOME
 ln -s $HOME/bin/.zshrc
